Move files to src
